@@ -11,42 +11,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.3",
-    "@angular/cdk": "^10.2.7",
-    "@angular/common": "~10.0.3",
-    "@angular/compiler": "~10.0.3",
-    "@angular/core": "~10.0.3",
+    "@angular/animations": "~11.1.1",
+    "@angular/cdk": "^11.1.1",
+    "@angular/common": "~11.1.1",
+    "@angular/compiler": "~11.1.1",
+    "@angular/core": "~11.1.1",
     "@angular/flex-layout": "^11.0.0-beta.33",
-    "@angular/forms": "~10.0.3",
-    "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "~10.0.3",
-    "@angular/platform-browser-dynamic": "~10.0.3",
-    "@angular/router": "~10.0.3",
-    "hammerjs": "^2.0.8",
-    "ngx-material-file-input": "^2.1.1",
-    "ngx-toastr": "^13.1.0",
-    "rxjs": "~6.5.5",
+    "@angular/forms": "~11.1.1",
+    "@angular/material": "^11.1.1",
+    "@angular/material-moment-adapter": "^11.2.3",
+    "@angular/platform-browser": "~11.1.1",
+    "@angular/platform-browser-dynamic": "~11.1.1",
+    "@angular/router": "~11.1.1",
+    "moment": "^2.29.1",
+    "ngx-currency": "^2.5.2",
+    "ngx-toastr": "^13.2.0",
+    "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.2",
-    "@angular/cli": "~10.0.2",
-    "@angular/compiler-cli": "~10.0.3",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "~0.1101.2",
+    "@angular/cli": "~11.1.2",
+    "@angular/compiler-cli": "~11.1.1",
+    "@angular/localize": "^11.1.1",
+    "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~5.2.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "typescript": "~4.1.2"
   }
 }
