@@ -5,11 +5,12 @@ const Local = mongoose.model('Local');
 const localCreate = (req, res) => {
     console.log('req.body.nombreDelLocal', req.body.nombreDelLocal);
     
-    if(!req.body.nombreDelLocal || !req.body.ruc || !req.body.direccionDelLocal || !req.body.nombreDelContacto || !req.body.emailDelContacto || !req.body.celularDelContacto){
+    if(!req.body.nombreDelLocal || !req.body.ruc || !req.body.direccionDelLocal || !req.body.nombreDelContacto || 
+        !req.body.emailDelContacto || !req.body.celularDelContacto || !req.body.urlMenuDigital  || !req.body.localNameUrl){
         console.log('Faltan datos en el request');
         return res
             .status(400)
-            .json({"message": "Faltan completar los campos!"});
+            .json({"message": "No se enviaron todos los parametros para el alta"});
 
     }
     Local.create({
@@ -19,6 +20,8 @@ const localCreate = (req, res) => {
         nombreDelContacto: req.body.nombreDelContacto,
         emailDelContacto: req.body.emailDelContacto,
         celularDelContacto: req.body.celularDelContacto,
+        urlMenuDigital: req.body.urlMenuDigital,
+        localNameUrl: req.body.localNameUrl,
         user: req.payload._id
     }, (err, local) => {
         if(err){
@@ -117,8 +120,33 @@ const localReadOneByEmail = (req, res) => {
             res.status(200).json(local);
         })
 }
+
+const localReadAllByUrlMenu = (req, res) => {
+    console.log('Parametro enviado a localUrlMenuSugerido:', req.params.urlmenu);
+    if(!req.params.urlmenu){
+        return res
+                .status(404)
+                .json({
+                    "message": "No se pudo encontrar el parametro urlmenu"
+                });
+    }
+    Local
+        .find({urlMenuDigital:req.params.urlmenu})
+        //.select() Agregar parametros al query
+        .exec((err, locales) => {
+            if (err){
+                return res
+                .status(404)
+                .json(err);
+            }
+            console.log('Se encontró el Local:',JSON.stringify(locales));
+            res.status(200).json(locales);
+        });
+        
+    
+}
 const localUpdateOne = (req, res) => {
-    console.log('Parametro enviado:', req.params.localid);
+    console.log('Update, Parametro localid enviado:', req.params.localid);
     if(!req.params.localid){
         return res
                 .status(404)
@@ -126,6 +154,15 @@ const localUpdateOne = (req, res) => {
                     "message": "No se pudo encontrar el Local, falta el localId"
                 });
     }
+    if(!req.body.nombreDelLocal || !req.body.ruc || !req.body.direccionDelLocal || !req.body.nombreDelContacto || 
+        !req.body.emailDelContacto || !req.body.celularDelContacto || !req.body.urlMenuDigital  || !req.body.localNameUrl){
+        return res
+                .status(404)
+                .json({
+                    "message": "No se enviaron todos los parametros de actualización"
+                });
+    }
+
     Local
         .findById(req.params.localid)
         //.select() Agregar parametros al query
@@ -141,18 +178,22 @@ const localUpdateOne = (req, res) => {
                 .status(404)
                 .json(err);
             }
+            console.log('Local encontrado:', local);
             local.nombreDelLocal = req.body.nombreDelLocal;
             local.ruc = req.body.ruc;
             local.direccionDelLocal = req.body.direccionDelLocal;
             local.nombreDelContacto = req.body.nombreDelContacto;
             local.emailDelContacto = req.body.emailDelContacto;
             local.celularDelContacto = req.body.celularDelContacto;
+            local.urlMenuDigital = req.body.urlMenuDigital;
+            local.localNameUrl = req.body.localNameUrl;
             local.save((err, localUpdated) =>{
                 if(err){
                     res
                         .status(404)
                         .json(err);
                 } else{
+                    console.log('Local actualizado con exito!');
                     res
                         .status(200)
                         .json(localUpdated);
@@ -167,6 +208,7 @@ module.exports = {
     localCreate,
     localReadOne,
     localUpdateOne,
-    localReadByUserId
+    localReadByUserId,
+    localReadAllByUrlMenu
     
 }
