@@ -1,15 +1,10 @@
 import { Component } from '@angular/core';
-import { AuthenticationFacade } from './authentication/services/authentication.facade';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
   title = 'menuDigital';
-  
-  constructor(){
-   
-  }
 }
