@@ -1,47 +1,28 @@
+import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { LOCALE_ID,NgModule } from '@angular/core';
-
+import {AuthenticationModule} from './authentication/authentication.module';
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import {CoreModule} from '../app/core/core.module';
-import {AuthenticationModule} from '../app/authentication/authentication.module';
-import {AppRoutingModule} from './app-routing.module';
-import {SharedModule} from './shared/shared.module';
 import { ToastrModule } from 'ngx-toastr';
-import { MatDialogRef } from '@angular/material/dialog';
-import { registerLocaleData } from '@angular/common';
-import localePy from '@angular/common/locales/es-PY';
-
-registerLocaleData(localePy, 'es');
+import { CoreModule } from './core/core.module';
 
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    
   ],
-  
   imports: [
     BrowserModule,
-    BrowserAnimationsModule,
-    CoreModule,
-    AuthenticationModule,
     AppRoutingModule,
-    SharedModule,
-    ToastrModule.forRoot({
-      timeOut: 10000,
-      positionClass: 'toast-top-center',
-      preventDuplicates: true,
-    }),
+    BrowserAnimationsModule,
+    AuthenticationModule,
+    CoreModule,
+    ToastrModule.forRoot()
     
-
   ],
-  providers: [
-    {
-      provide: MatDialogRef,
-      useValue: {}
-    },
-    { provide: LOCALE_ID, useValue: 'es-PY' }
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
