@@ -1,22 +1,22 @@
-import { Component, EventEmitter, OnInit, Output } from '@angular/core';
-import {FormBuilder, FormControl, FormGroup, Validators} from '@angular/forms';
+import { Component, EventEmitter, OnInit, Output, ViewChild, ElementRef } from '@angular/core';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { MatDialogRef } from '@angular/material/dialog';
 
 @Component({
-  selector: 'app-login',
+  selector: 'login-app',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss']
 })
 export class LoginComponent implements OnInit {
-  @Output() private sendLoginForm = new EventEmitter<any>();
+  @Output() sendLoginForm = new EventEmitter<any>();
   form: FormGroup;
-  private flatlogicEmail = 'admin@flatlogic.com';
-  private flatlogicPassword = 'admin';
+  
 
-  constructor(private readonly formBuilder:FormBuilder) {
+  constructor(private readonly formBuilder:FormBuilder, public dialogRef: MatDialogRef<LoginComponent>) {
     this.form = this.formBuilder.group(
       {
-        email: new FormControl(this.flatlogicEmail, [Validators.required, Validators.email]),
-        password: new FormControl(this.flatlogicPassword, [Validators.required])
+        email: new FormControl('', [Validators.required, Validators.email]),
+        password: new FormControl('', [Validators.required])
       });
    }
 
@@ -24,10 +24,16 @@ export class LoginComponent implements OnInit {
   }
 
   
+  //Cerrar dialogo de Nuevo MenuItem
+  public onNoClick(){
+    this.dialogRef.close();
+  }
   public login(): void {
     //console.log('LoginComponent.login()');
+    this.dialogRef.close();
     if (this.form.valid) {
       this.sendLoginForm.emit(this.form.value);
     }
   }
+
 }
