@@ -4,6 +4,6 @@ export interface User{
     name?: string;
     password?: string;
     token?: string;
-    profile?: string;
+    role?: string;
 
 }
