@@ -1,16 +1,19 @@
-import {Injectable} from '@angular/core';
-import {Subject, BehaviorSubject, Observable} from 'rxjs';
-import {Local} from '../models/local-model';
+import { Injectable } from "@angular/core";
+import { BehaviorSubject, Observable } from "rxjs";
+import { Local } from "../models/local.model";
 
 @Injectable()
 export class LocalStateService{
-   private currentLocal:Local=null;
+    private _local: BehaviorSubject<Local> = new BehaviorSubject<Local>(null);
 
-   public setCurrentLocal(newLocal:Local){
-       this.currentLocal = newLocal;
-   }
-    
-   public getCurrentLocal(){
-       return this.currentLocal;
-   }
+    public readonly local$:Observable<Local> = this._local.asObservable();
+
+    constructor(){}
+
+    public setLocal(localParam:Local){
+        console.log('LocalState:', localParam);
+        this._local.next(localParam);
+        
+    }
+
 }
