@@ -1,18 +1,16 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { MainMenuComponent } from './container/main-menu/main-menu.component';
 
-import { MainMenuListComponent } from './container/main-menu/main-menu-list.component';
-
-
-const routes: Routes = [
-  {
-    path: '',
-    component: MainMenuListComponent
-  }
-];
+const routes:Routes = [
+    {
+        path: '',
+        component: MainMenuComponent
+    }
+]
 
 @NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
+  })
 export class PublicMenuRoutingModule { }
