@@ -1,21 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MainMenuListComponent } from './container/main-menu/main-menu-list.component';
+import { MainMenuComponent } from './container/main-menu/main-menu.component';
 import { PublicMenuRoutingModule } from './public-menu-routing.module';
 import { SharedModule } from '../shared/shared.module';
-import { MainMenuFacade } from './services/mainMenu.facade';
-import { PublicMenuItemComponent } from './components/public-menu-item/public-menu-item.component';
-import { PersonalizarMenuItemComponent } from './components/personalizar-menu-item/personalizar-menu-item.component';
+import { PublicMenuFacade } from './services/publicMenuFacade.service';
 
 
 
 @NgModule({
-  declarations: [MainMenuListComponent, PublicMenuItemComponent, PersonalizarMenuItemComponent],
+  declarations: [MainMenuComponent],
   imports: [
     CommonModule,
     PublicMenuRoutingModule,
     SharedModule
   ],
-  providers:[MainMenuFacade]
+  providers:[
+    PublicMenuFacade
+  ]
 })
 export class PublicMenuModule { }
