@@ -1,8 +1,6 @@
-import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
-import {Observable} from 'rxjs';
-import {User} from '../../../core/models/user.model';
-import {AuthenticationFacade} from '../../../authentication/services/authentication.facade';
-import { Router } from '@angular/router';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { Observable } from 'rxjs';
+import { User } from 'src/app/core/models/user.model';
 
 @Component({
   selector: 'app-header',
@@ -11,29 +9,25 @@ import { Router } from '@angular/router';
 })
 export class HeaderComponent implements OnInit {
 
-  @Input() isMenuOpened:boolean;
-  @Output() isShowSidebar = new EventEmitter<boolean>();
-  public user: User;
-
-  constructor(
-    private authFacade: AuthenticationFacade,
-    private router: Router
-
-  ) { }
+  @Input() userName$: Observable<string>;
+  @Output() signOutEvent = new EventEmitter<string>();
+  @Output() menuClick: EventEmitter<boolean> = new EventEmitter();
+  toggleValue: boolean=true;
+  constructor() { }
 
   ngOnInit(): void {
-    this.user = this.authFacade.getCurrentUser();
-    //console.log('HeaderComponent.user',this.user);
   }
 
-  public openMenu(): void {
-    this.isMenuOpened = !this.isMenuOpened;
-
-    this.isShowSidebar.emit(this.isMenuOpened);
-  }
   public signOut(): void {
-    this.authFacade.signOut();
-
-    this.router.navigate(['login']);
+    console.log('Sign Out');
+    this.signOutEvent.emit('SignOut');
+    
   }
+
+  public toggleForMenuClick(){
+    this.toggleValue = ! this.toggleValue;
+    this.menuClick.emit(this.toggleValue);
+    
+  }
+
 }
